% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scRCD}
\alias{scRCD}
\title{scRCD main function: to calculate rareness scores of cells for scRNA-seq data}
\usage{
scRCD(pc, anchor.frac = 0.15, alpha = 0.975, t = 10, n.cores = 10)
}
\arguments{
\item{pc}{The top PCs selected from PCA or other dimension reduction methods,
the columns refer to features and rows refer to samples}

\item{anchor.frac}{The proportion of anchor cells with respect to all cells}

\item{alpha}{The percentage of quantile for the distance threshold for bipartite construction.
The distance threshold is chosen by the alpha-quantile of the Euclidean distances between cells.
The distance threshold is to control the bipartite distance between cells and anchor cells
such that only the edges of anchor cells with small distances will remain.
While constructing the bipartite and calculating the distance of each cell to anchor cells,
the anchor cells with the Euclidean distance (of PCs) from the cell lower than the threshold are selected and the bipartite distances
will be calculated by similar to k-NN graphs, the bipartite distance between the cell and the remaining anchor cells will be set to 0.}

\item{t}{the number of total repetitions to avoid the random effects of uniformly sampling anchor cells}

\item{n.cores}{the number of cores called for computation, 10 by default}
}
\value{
the rareness scores for all the cell samples, the high rareness score of a cell suggest
the "further" this cell is compared to majority cells and a high possibility of this cell to be rare
}
\description{
Sorting the cells by the rareness scores, which are calculated from a graphical model via this function.
First, the anchor cells are randomly selected uniformly from all cells.
The bipartite (between anchor cells and all the cells) is constructed.
Then the affinity matrix is calculated as the bipartite multiply its transpose.
Finally the rareness scores are computed from a spectrum-clustering based optimization problem.
This steps are repeated until maximal repetition numbers achieved to avoid the effects of randomly sampling anchor cells.
The final scores are the averages of repeatly calculated scores.
For each repetition, the scores are calculated by calling rare.cell.scores()
}
