% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{scRCD.visualize}
\alias{scRCD.visualize}
\title{t-SNE visualization of the rare cells}
\usage{
scRCD.visualize(tsne, scores, cell.type, frac.list = NULL, frac = NULL)
}
\arguments{
\item{tsne}{the tsne object returned by Rtsne::Rtsne()}

\item{scores}{the calculated rareness scores by scRCD or other methods,
please remind that the higher rareness scores suggest higher possibility of being rare}

\item{cell.type}{a one-colunm matrix indicating the cell labels (by clustering or biomarkers) for each cell}

\item{frac.list}{a vector of desired fractions of rare cells for visualizations}

\item{frac}{a numeric value of the desired fraction, either frac.list or frac should be input}
}
\value{
A list object containing the x and y of tSNE, and the ggplot objects containing different visualizations depending on input arguments.
 If input a vector of fractions, returns the t-SNE visualization plots for each input rare cell fraction
 as well as the t-SNE plot for all the cells (cell types labelled in different colors);
 if input a single fraction value, the tSNE plot of rare cells at the single fraction,
 the visualization of uniformly sampled cells as the input fraction, and the visualization of all the cells.
}
\description{
Given the fractions of desired rare cells, this function visualize the determined rare cells by t-SNE compared to all cells in data set
}
